Big changes  on handling files
If an user tries to spend, clear expense or check expense on a unexisting file it  won't throw an error, but a warning that the file or category does  not exist, and the program will continue running.

Also, prints formatted strings up tp2  decimal spaces and improved print_stats() function with a new sub function called check expenses.

+++ budget_control.py
@@ -15,6 +15,15 @@ def print_lista(lista):
         print(x)
 
 
+def msv(k):
+    mf = '{:.2f}'.format(k)
+    dec = mf[-2:]
+    intf = '{:,}'.format(int(k))
+    antes = str(intf).replace(',', '.')
+    fn = '{}{}{}'.format(antes, ',', dec)
+    return fn
+
+
 def ppv(x):
     nx = x.replace(',', '.')
     if nx.count('.') == 0:
@@ -30,17 +39,9 @@ def formatação(m):
             v = ppv(str(input(m)))
             return v
         except ValueError:
-            print('\nYou  typed one or more invalid characters. '
-                  'Type only numbers, splitting the decimals with either dots or commas!\n')
+            print('\nVocê digitou um ou mais caracteres inválidos. '
+                  'Digite apenas números, separando as casas decimais com ponto ou vírgula!\n')
             continue
-
-
-def try_cat(category):
-    if ' ' in category:
-        category = category.replace(' ', '_')
-    else:
-        pass
-    return category
 
 
 def sum_files(_type):
@@ -60,7 +61,7 @@ def sum_files(_type):
 
             else:
                 with open(name) as f:
-                    number = float(f.read())
+                    number = int(f.read())
                     number_list.append(number)
                     sum_number_list = sum(number_list)
 
@@ -84,19 +85,45 @@ def get_file_name(category, _type):
 
 
 def action_file(file, mode='r', amount=0):
-    if mode == 'w':
-        with open(file, mode) as f:
-            value = float(f.write(amount))
-            return value
+    try:
+        if mode == 'w':
+            with open(file, mode) as f:
+                value = float(f.write(amount))
+                return value
 
-    elif mode == 'a':
-        with open(file, mode) as f:
-            value = f.write(amount)
+        elif mode == 'a':
+            with open(file, mode) as f:
+                value = f.write(amount)
 
-    elif mode == 'r' and amount == 0:
-        with open(file) as f:
-            value = f.read()
-            return value
+        elif mode == 'r' and amount == 0:
+            with open(file) as f:
+                value = f.read()
+                return value
+    except IOError as err:
+        print('Could not get file. Be sure it exists')
+        # errno.ENOENT
+        '''print(err.errno)
+                                print(err.strerror)'''
+
+
+def try_file(name, type_):
+    x = action_file(get_file_name(name, type_))
+    if x is None:
+        return False
+    else:
+        return True
+
+
+def file_not_found():
+    print('The category does not exist.')
+
+
+def try_cat(category):
+    if ' ' in category:
+        category = category.replace(' ', '_')
+    else:
+        pass
+    return category
 
 
 def get_budget():
@@ -106,13 +133,14 @@ def get_budget():
 
 def add_money(amount):
     action_file(get_file_name('budget', 'data'), mode='w', amount=str(amount))
-    print('\nYour new balance is: {}\n'.format(amount))
+    print('\nYour new balance is: {}\n'.format(msv(amount)))
 
 
 def set_limit(category, amount):
     budget = get_budget()
     total_limits = sum_files('limit')
     soma = amount + total_limits
+    category = try_cat(category)
     filepath_categorias = get_file_name('categorias', 'data')
     contents = action_file(filepath_categorias)
 
@@ -121,17 +149,17 @@ def set_limit(category, amount):
 
     elif soma > budget and category not in contents:
         sobra = budget - total_limits
-        print('\nThe limit exceeded your budget. \nYou have ${} left on your budget\n'.format(sobra))
+        print('\nThe limit exceeded your budget. \nYou have ${} left on your budget\n'.format(msv(sobra)))
 
     else:
-        category = try_cat(category)        
+
         if category not in contents:
             action_file(filepath_categorias, mode='a', amount=(category + ' '))
 
         filepath = get_file_name(try_cat(category), 'limit')
         action_file(filepath, mode='w', amount=str(amount))
 
-        print('\nYour new limit on {} is: {} \n'.format(category, amount))
+        print('\nYour new limit on {} is: {} \n'.format(category, msv(amount)))
 
         try:
             default_file = os.path.join(os.getcwd(), 'data', 'limit', 'default_limit.txt')
@@ -145,24 +173,33 @@ def spend(category, amount):
     category = try_cat(category)
     filepath_limit = get_file_name(category, 'limit')
     filepath_spend = get_file_name(category, 'spend')
+    contents_c = action_file(get_file_name('categorias', 'data'))
 
-    if budget == 0:
-        print('\nYou have to register an income first! \nFor that you can use the command "add money" and type in the amount.\n')
+    if try_file(category, 'spend') is True:
 
-    action_file(filepath_spend, mode='a', amount=(str(amount) + '\n'))
-    limit = float(action_file(filepath_limit))
-    fs_contents = action_file(filepath_spend)
-    number_list = fs_contents.split()
-    number_list = list(map(float, number_list))
-    number = sum(number_list)
-    saldo = limit - number
-    print('\nYou just spent ${} on {}. \nYour limit is ${} \nYou stil have ${} on that category\n'.format(str(amount), category, str(limit), str(saldo)))
+        if budget == 0:
+            print('\nYou have to register an income first! \nFor that you can use the command "add money" and type in the amount.\n')
 
-    try:
-        default_file = os.path.join(os.getcwd(), 'data', 'spend', 'default_spend.txt')
-        os.remove(default_file)
-    except:
-        pass
+        elif category not in contents_c:
+            print('There are no records on that category. Did you typed in correctly?')
+
+        else:
+            action_file(filepath_spend, mode='a', amount=(str(amount) + '\n'))
+            limit = float(action_file(filepath_limit))
+            fs_contents = action_file(filepath_spend)
+            number_list = fs_contents.split()
+            number_list = list(map(float, number_list))
+            number = sum(number_list)
+            saldo = limit - number
+            print('\nYou just spent ${} on {}. \nYour limit is ${} \nYou stil have ${} on that category\n'.format(str(amount), category, msv(limit), msv(saldo)))
+
+            try:
+                default_file = os.path.join(os.getcwd(), 'data', 'spend', 'default_spend.txt')
+                os.remove(default_file)
+            except:
+                pass
+    else:
+        file_not_found()
 
 
 def saldo():
@@ -172,8 +209,35 @@ def saldo():
 
 
 def clear_expense(category):
-    filepath = get_file_name(try_cat(category), 'spend')
-    action_file(filepath, mode='w', amount='')
+    if try_file(try_cat(category), 'spend') is True:
+        filepath = get_file_name(try_cat(category), 'spend')
+        action_file(filepath, mode='w', amount='')
+    else:
+        file_not_found()
+
+
+def check_expenses(category, print_=False):
+    if try_file(try_cat(category), 'spend') is True:
+        budget = get_budget()
+        filepath_limit = get_file_name(try_cat(category), 'limit')
+        filepath_spend = get_file_name(try_cat(category), 'spend')
+        limit = float(action_file(filepath_limit))
+        fs_contents = action_file(filepath_spend)
+        number_list = fs_contents.split()
+        number_list = list(map(float, number_list))
+        number = sum(number_list)
+        saldo = limit - number
+        contents_c = action_file(get_file_name('categorias', 'data'))
+        if category not in contents_c:
+            print('There are no records on that category. Did you typed in correctly?')
+
+        elif print_ is True:
+            print('Your limit in {} category is ${} \nYour total expenses on that category is ${}\nYou have ${} left on that category.\n'.format(category, msv(limit), msv(number), msv(saldo)))
+
+        else:
+            return number, saldo
+    else:
+        file_not_found()
 
 
 def print_stats():
@@ -188,15 +252,20 @@ def print_stats():
                 file_name = (name_list[len(name_list) - 1])
                 file_name = file_name.split('.')
                 file_name = file_name[0]
+                number, saldo = check_expenses(file_name)
                 file_name = file_name.replace('_', ' ')
-                print('Your limit in {} category is ${}'.format(file_name, limit))
+
+                if saldo == 0:
+                    print('Your limit in {} category is ${} \nYou have used all your limit on that category'.format(file_name, msv(limit)))
+                else:
+                    print('Your limit in {} category is ${} \nYour total expenses on that category is ${}\nYou have ${} left on that category.'.format(file_name, msv(limit), msv(number), msv(saldo)))
             print('\n')
 
         except IOError as exc:
             if exc.errno != errno.EISDIR:
                 raise
 
-    print('Your total limit is {} \nYour total spent is {}\n'.format(sum_files('limit'), sum_files('spend')))
+    print('Your total limit is {} \nYour total spent is {}\n'.format(msv(sum_files('limit')), msv(sum_files('spend'))))
 
 
 def prompt():
@@ -229,10 +298,14 @@ def prompt():
             clear_expense(category)
 
         elif prompt == 'balance':
-            print('\nYour current balance is: {}\n'.format(saldo()))
+            print('\nYour current balance is: {}\n'.format(msv(saldo())))
 
         elif prompt == 'check limits':
             print_stats()
+
+        elif prompt == 'check expenses':
+            category = input('What category would you like to check the expenses? \n')
+            check_expenses(category, print_=True)
 
         elif prompt == 'quit':
             break
